Enhance support for language id passing when document open
With the previous implementation when supporting for multiple language id
which are not same as the ext we need to define multiple language server
definitions which cause multiple language servers to start.

With new implementation we can support for multiple languages with one
server. This also enable proper language id passing when using
extension patterns.

Signed-off-by: gayanper <gayanper@gmail.com>