refactor: Drop the type arguments, making it easier to implement
Judging from the usage of the API, there is no need to use type
arguments. The type isn't provided (in form of a class instance) and
will be erased anyway during the compilation process.

The API called expecting some interfaces to be implemented, but that
can be expressed using plain old Java interfaces.

On the other side, having those generic type arguments make it difficult
to implement the interfaces, as the compiler tries to ensure it is
possible to cast to `T`, which however is completely unknown.